# openapi.yaml
openapi: 3.0.0
paths:
  /auth:
    post:
      operationId: createToken
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Authentication'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '400':
          description: Invalid payload
        '401':
          description: Invalid credentials
        '500':
          description: Something went wrong
      security:
        - bearerAuth: []
      summary: Create a token for a user
      tags: &ref_0
        - Authentication
  /auth/register:
    post:
      operationId: registerUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Registration'
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Invalid payload
        '409':
          description: Username already taken
        '500':
          description: Something went wrong
      summary: Register a new user
      tags: *ref_0
...
info:
  title: Aviobook Code Challenge
  description: CRUD API for the AVIOBOOK assessment
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    Authentication:
      type: object
      properties:
        username:
          type: string
          example: testuser
          description: The username of the user
        password:
          type: string
          example: pass1234
          description: The password of the user
      required:
        - username
        - password
    Registration:
      type: object
      properties:
        username:
          type: string
          example: newuser
          description: Unique username for the user
        password:
          type: string
          example: strongPassword123
          description: User password
      required:
        - username
        - password
...